Improve gradle structure (6aafd67500ea0196e24703a2d349f506948a7d83)
