+++ application/src/main/java/org/togetherjava/formatter/tokenizer/Lexer.java
