+++ application/src/main/java/org/togetherjava/formatter/tokenizer/Token.java
