+++ application/src/main/java/org/togetherjava/formatter/tokenizer/TokenType.java
