+++ application/src/main/java/org/togetherjava/tjbot/PingPongListener.java
