+++ application/src/main/java/org/togetherjava/tjbot/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Entry package for the TJ-Bot. See {@link org.togetherjava.tjbot.Application} to get started.
+ */
+package org.togetherjava.tjbot;
