+++ application/src/test/java/org/togetherjava/tjbot/ApplicationTest.java
@@ -3,9 +3,9 @@ package org.togetherjava.tjbot;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public final class MainTest {
-	@Test
-	void testFoo() {
-		Assertions.assertTrue(true);
-	}
+public final class ApplicationTest {
+    @Test
+    void testFoo() {
+        Assertions.assertTrue(true);
+    }
 }
