+++ build.gradle
@@ -1,34 +1,13 @@
 plugins {
     id 'java'
     id "com.diffplug.spotless" version "5.15.0"
+    id "name.remal.sonarlint" version "1.5.0"
 }
-
-group 'org.togetherjava'
-version '1.0-SNAPSHOT'
 
 sourceCompatibility = JavaVersion.VERSION_16
 targetCompatibility = JavaVersion.VERSION_16
 
-repositories {
-    mavenCentral()
-    maven {
-        name 'm2-dv8tion'
-        url 'https://m2.dv8tion.net/releases'
-    }
-}
-
-dependencies {
-    implementation 'net.dv8tion:JDA:4.3.0_324'
-    implementation 'org.apache.logging.log4j:log4j-api:2.14.1'
-    implementation 'org.apache.logging.log4j:log4j-core:2.14.1'
-    implementation 'org.apache.logging.log4j:log4j-slf4j18-impl:2.14.1'
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.2'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.2'
-}
-
-test {
-    useJUnitPlatform()
-}
+compileJava.dependsOn 'spotlessApply'
 
 spotless {
     java {
