+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'TJ-Bot'
-
+include('application')
