--- src/main/java/org/togetherjava/tjbot/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Entry package for the TJ-Bot. See {@link org.togetherjava.tjbot.Main} to get started.
- */
-package org.togetherjava.tjbot;
